requirements.txt
streamlit==1.28.1
PyMuPDF==1.23.7
pdfplumber==0.10.3
openai==1.12.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
tiktoken==0.5.2